Refactor codebase and separate library from CLI
This commit significantly refactors the codebase. This enables nb-clean
to be used as a Python library in addition to as a command line tool,
and introduces unit tests for both.

When using the nb-clean CLI, it is no longer required to pass flags to
indicate input and output files, as nb-clean will operate on all files
passed as positional arguments in place. If no positional arguments are
passed, the input notebook is read from standard input and, for the
check subcommand, the output notebook is written to standard output.

To support this and improve usability, the command line interface has
changed. The new interface is documented in README.md.