Bump required nbformat version to 5.0.5
Versions earlier than 5.0.5 of `nbformat` run into file reading
problems when passing a filename to `nb-clean clean`, as path-like
objects are not supported.  That feature was added in
https://github.com/jupyter/nbformat/pull/173

Bumping the version fixes this issue for installing with `pip`.
Installs with `poetry` didn't experience this as `poetry.lock` already
contains a newer version which has this feature.