Create reserved_metadata_tags.ipynb
Some metadata tags are conventionally defined and may be used to modify the presentation of cells or add meaningful cell "tags".

As such, in some circumstances such metadata may be considered as notebook content that should be persisted rather than being summarily deleted from cell metadata when a notebook is cleaned.