Document corner case parsing cell metadata fields in CLI
Because `--preserve-cell-metadata` can take zero or more arguments,
the argument parser assigns any positional arguments following this
flag as fields to preserve rather than notebook paths to process.
This means if you're not passing any custom fields to preserve, the
`--preserve-cell-metadata` flag needs to appear after the paths to
be processed, or the notebook path should be preceded with `--` to
tell the argument parser it's not an argument to
`--preserve-cell-metadata`.